+++ moviu_server/config.py
@@ -10,7 +10,7 @@ from typing import Optional
 
 CONFIG_DIR = Path.home() / ".moviu_printer"
 CONFIG_FILE = CONFIG_DIR / "config.json"
-VERSION = "1.4.2"
+VERSION = "1.4.3"
 
 
 @dataclass